requirements.txt
streamlit
python-dotenv
langchain
langchain-community
PyPDF2
google-generativeai
chromadb
langchain_google_genai
faiss-cpu